requirements.txt
Cython==0.23
Django==1.9.5
Kivy==1.9.1
Kivy-Garden==0.1.4
MySQL-python==1.2.5
Pygments==2.1.3
argparse==1.2.1
djangorestframework==3.3.3
requests==2.9.1
wsgiref==0.1.2
